requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
simpleaudio>=1.0.4
numpy>=1.24.0

# Testing and monitoring
psutil>=5.9.0
